refactored TagCloudWordRenderer
