+++ TagCloud/WordCloudLayouter/IWordCloudLayouter.cs
@@ -1,0 +1,8 @@
+using System.Drawing;
+
+namespace TagCloud.WordCloudLayouter;
+
+public interface IWordCloudLayouter
+{
+    public void DrawWordCloud(Graphics graphics);
+}
